TASK: Rename scss files to css
