+++ packages/react-ui-components/src/Badge/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import Badge from './badge';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.badge, style)(Badge);
