+++ packages/react-ui-components/src/Bar/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import Bar from './bar';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.bar, style)(Bar);
