+++ packages/react-ui-components/src/Button/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import Button from './button';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.button, style)(Button);
