+++ packages/react-ui-components/src/CheckBox/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 
 import identifiers from '../identifiers';
 import CheckBox from './checkBox';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.checkBox, style)(CheckBox);
