+++ packages/react-ui-components/src/DateInput/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 
 import identifiers from '../identifiers';
 import DateInput from './dateInput';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.dateInput, style)(DateInput);
