+++ packages/react-ui-components/src/Dialog/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 
 import identifiers from '../identifiers';
 import Dialog from './dialog';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.dialog, style)(Dialog);
