+++ packages/react-ui-components/src/Headline/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import Headline from './headline';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.headline, style)(Headline);
