+++ packages/react-ui-components/src/Icon/index.ts
@@ -3,6 +3,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 import Icon from './icon';
 
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.icon, style)(Icon);
