+++ packages/react-ui-components/src/IconButton/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import IconButton from './iconButton';
-import style from  './style.scss';
+import style from  './style.css';
 
 export default themr(identifiers.iconButton, style)(IconButton);
