+++ packages/react-ui-components/src/IconButtonDropDown/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import IconButtonDropDown from './iconButtonDropDown';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.iconButtonDropDown, style)(IconButtonDropDown);
