+++ packages/react-ui-components/src/Label/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import Label from './label';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.label, style)(Label);
