+++ packages/react-ui-components/src/ListPreviewElement/index.js
@@ -1,6 +1,6 @@
 import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
-import style from './style.scss';
+import style from './style.css';
 import ListPreviewElement from './listPreviewElement';
 
 const ThemedListPreviewElement = themr(identifiers.listPreviewElement, style)(ListPreviewElement);
