+++ packages/react-ui-components/src/MultiSelectBox/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable camelcase, react/jsx-pascal-case */
 import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
-import style from './style.scss';
+import style from './style.css';
 import MultiSelectBox from './multiSelectBox';
 
 const ThemedMultiSelectBox = themr(identifiers.multiSelectBox, style)(MultiSelectBox);
