+++ packages/react-ui-components/src/SideBar/index.ts
@@ -2,6 +2,6 @@ import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
 
 import SideBar from './sideBar';
-import style from './style.scss';
+import style from './style.css';
 
 export default themr(identifiers.sideBar, style)(SideBar);
