+++ packages/react-ui-components/src/Tabs/index.ts
@@ -1,6 +1,6 @@
 import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
-import style from './style.scss';
+import style from './style.css';
 import Tabs from './tabs';
 
 export default themr(identifiers.tabs, style)(Tabs);
