+++ packages/react-ui-components/src/TextArea/index.ts
@@ -1,6 +1,6 @@
 import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
-import style from './style.scss';
+import style from './style.css';
 import TextArea from './textArea';
 
 const ThemedTextArea = themr(identifiers.textArea, style)(TextArea);
