+++ packages/react-ui-components/src/ToggablePanel/index.js
@@ -1,6 +1,6 @@
 import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
-import style from './style.scss';
+import style from './style.css';
 import ToggablePanel, {
     Header,
     Contents
