+++ packages/react-ui-components/src/ToggablePanel/style.css
