+++ packages/react-ui-components/src/Tree/index.js
@@ -1,7 +1,7 @@
 import {themr} from '@friendsofreactjs/react-css-themr';
 import identifiers from '../identifiers';
-import treeStyle from './tree.scss';
-import nodeStyle from './node.scss';
+import treeStyle from './tree.css';
+import nodeStyle from './node.css';
 import dragLayerStyle from './dragLayer.css';
 import Tree from './tree';
 import Node, {
